refactor: 카드 컴포넌트 분리 및 리액트 파일 jsx로 변경 완료
